feat: modularize the control layer (buffer,picker,window) (#4)
* chore: stylua

* feat: add buffer.lua

Module to manage buffer with names and IDs.

* feat: add window.lua

Module to manager windows with create a buffer or without.

* feat: add picker.lua

Module to interact with the window and buffer module. Like... a... picker.

* feat: add utils.lua

Utils module. I don't have anything to say about this module for now.

* feat: create cocoon main module using picker to instance a window

Build an init.lua to pass the submodule to the top layer.

* update README.org

* chore: stylua again